[rename]: Product 디렉터리 파일 구조 변경(#82)
