+++ src/layouts/Product/DetailBottom/ProductItem.module.scss
