+++ src/layouts/Product/DetailBottom/index.module.scss
