+++ src/layouts/Product/DetailContent/DetailContentAccordion.module.scss
