+++ src/layouts/Product/DetailContent/DetailContentAccordion.tsx
