+++ src/layouts/Product/DetailContent/index.module.scss
