+++ src/layouts/Product/DetailContent/index.tsx
