+++ src/layouts/Product/DetailMain/ProductDetailsCarousel.module.scss
