+++ src/layouts/Product/DetailMain/ProductDetailsCarousel.tsx
