+++ src/layouts/Product/DetailMain/index.module.scss
