+++ src/pages/Product/index.module.scss
@@ -10,6 +10,5 @@
 
   .area_buy_info {
     width: 30%;
-    // background-color: antiquewhite;
   }
 }
